+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [next]
+## [0.7.0]
 * BREAKING CHANGE: improvement in animations to SimplePlayer and SimpleEnemy. Now use `SimpleDirectionAnimation` to manipulate animations.
 * add `GameColorFilter` it is now possible to add color filter in the game.
 * Possible to load maps made by Tiled by the server. Just pass the link as path.